# docs/swagger.yaml
definitions:
...
  requests.CreateBannerRequest:
    properties:
      end_date:
        type: string
      end_time:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
      start_time:
        type: string
    required:
    - end_date
    - end_time
    - name
    - start_date
    - start_time
    type: object
...
  requests.UpdateBannerRequest:
    properties:
      banner_id:
        type: integer
      end_date:
        type: string
      end_time:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
      start_time:
        type: string
    required:
    - banner_id
    - end_date
    - end_time
    - name
    - start_date
    - start_time
    type: object
...
  response.ApiResponseBanner:
    properties:
      data:
        $ref: '#/definitions/response.BannerResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseBannerAll:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseBannerDelete:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseBannerDeleteAt:
    properties:
      data:
        $ref: '#/definitions/response.BannerResponseDeleteAt'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponsePaginationBanner:
    properties:
      data:
        items:
          $ref: '#/definitions/response.BannerResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationBannerDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.BannerResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.BannerResponse:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      end_time:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
    type: object
  response.BannerResponseDeleteAt:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      end_date:
        type: string
      end_time:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/banner:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all banners
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of banners
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationBanner'
        "500":
          description: Failed to retrieve banner data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all banners
      tags:
      - Banner
  /api/banner/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a banner by ID
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Banner data
          schema:
            $ref: '#/definitions/response.ApiResponseBanner'
        "400":
          description: Invalid banner ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve banner data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find banner by ID
      tags:
      - Banner
  /api/banner/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active banners
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active banners
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationBannerDeleteAt'
        "500":
          description: Failed to retrieve banner data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active banners
      tags:
      - Banner
  /api/banner/create:
    post:
      consumes:
      - application/json
      description: Create a new banner with the provided details
      parameters:
      - description: Create banner request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateBannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created banner
          schema:
            $ref: '#/definitions/response.ApiResponseBanner'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create banner
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new banner
      tags:
      - Banner
  /api/banner/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a Banner record by its ID.
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted Banner record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseBannerDelete'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 'Failed to delete Banner:'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete a Banner
      tags:
      - Banner
  /api/banner/delete/all:
    post:
      consumes:
      - application/json
      description: Permanently delete a banner record by its ID.
      parameters:
      - description: banner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted banner record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseBannerAll'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 'Failed to delete banner:'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete a banner
      tags:
      - Banner
  /api/banner/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a trashed Banner record by its ID.
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored Banner
          schema:
            $ref: '#/definitions/response.ApiResponseBannerDeleteAt'
        "400":
          description: Invalid Banner ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore Banner
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed Banner
      tags:
      - Banner
  /api/banner/restore/all:
    post:
      consumes:
      - application/json
      description: Restore a trashed Banner record by its ID.
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored Banner all
          schema:
            $ref: '#/definitions/response.ApiResponseBannerAll'
        "400":
          description: Invalid Banner ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore Banner
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed Banner
      tags:
      - Banner
  /api/banner/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed banners
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active banners
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationBannerDeleteAt'
        "500":
          description: Failed to retrieve banner data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed banners
      tags:
      - Banner
  /api/banner/trashed/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a trashed Banner record by its ID.
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved trashed Banner
          schema:
            $ref: '#/definitions/response.ApiResponseBannerDeleteAt'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve trashed Banner
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve a trashed Banner
      tags:
      - Banner
  /api/banner/update/{id}:
    post:
      consumes:
      - application/json
      description: Update an existing banner record with the provided details
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update banner request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateBannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated banner
          schema:
            $ref: '#/definitions/response.ApiResponseBanner'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to update banner
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update an existing banner
      tags:
      - Banner
...
swagger: "2.0"